+++ tests/QueryType/Select/RequestBuilderTest.php
@@ -93,6 +93,25 @@ class RequestBuilderTest extends TestCase
         );
     }
 
+    public function testSelectUrlWithCachedFilters()
+    {
+        $this->query->addFilterQuery(
+            new FilterQuery(['key' => 'f1', 'query' => 'published:true', 'local_cache' => true, 'local_cost' => 95])
+        );
+        $this->query->addFilterQuery(
+            new FilterQuery(['key' => 'f2', 'local_tag' => ['t1', 't2'], 'query' => 'category:23', 'local_cache' => false])
+        );
+        $request = $this->builder->build($this->query);
+
+        $this->assertNull($request->getRawData());
+
+        $this->assertSame(
+            'select?omitHeader=true&wt=json&json.nl=flat&q=*:*&start=0&rows=10&fl=*,score'.
+            '&fq={!cache=true cost=95}published:true&fq={!tag=t1,t2 cache=false}category:23',
+            urldecode($request->getUri())
+        );
+    }
+
     public function testWithComponentNoBuilder()
     {
         $this->builder->build($this->query);