+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the solarium library will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- setting limit for pivot facets
+
 ## [5.1.5]
 ### Security
 - Remove explicit requirements for symfony/cache because of CVE-2019-18889