+++ CHANGELOG.md
@@ -4,14 +4,17 @@ All notable changes to the solarium library will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [5.1.6]
 ### Added
 - Range facet pivot support
 - Support for useConfiguredElevatedOrder
 - FilterQuery::setCache and FilterQuery::setCost()
 
 ### Fixed
-- setting limit for pivot facets
+- Setting limit for pivot facets
+
+### Changed
+- Internal handling of Solr local parameters
 
 ### Deprecated
 - Helper::cacheControl(). Use FilterQuery::setCache() and FilterQuery::setCost() instead.